refactor merge-tables by table
